Relocate feature manager
